Create wallet infrastructure.
This diff adds wallet infrastructure that can be used to create custom
bitcoin wallets depending on its needs.
