+++ bitcoin/bitcoin_test.go
@@ -5,33 +5,9 @@
 package bitcoin
 
 import (
-	"bytes"
-	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"errors"
-	"fmt"
 	"testing"
-	"time"
-
-	"github.com/btcsuite/btcd/btcutil"
-	"github.com/btcsuite/btcd/btcutil/hdkeychain"
-	"github.com/btcsuite/btcd/chaincfg"
-	"github.com/btcsuite/btcd/txscript"
-	"github.com/davecgh/go-spew/spew"
-	"github.com/juju/loggo"
-	"github.com/tyler-smith/go-bip39"
-
-	"github.com/hemilabs/heminetwork/api/tbcapi"
-	"github.com/hemilabs/heminetwork/hemi"
-	"github.com/hemilabs/heminetwork/service/tbc"
-	"github.com/hemilabs/heminetwork/service/tbc/peer/rawpeer"
 )
-
-func digest256(x []byte) []byte {
-	xx := sha256.Sum256(x)
-	return xx[:]
-}
 
 func TestValidateMerklePathOne(t *testing.T) {
 	index := 3
@@ -278,402 +254,4 @@ func TestCheckMerkleChain(t *testing.T) {
 			t.Errorf("Test %d - failed to validate merkle chain: %v", i, err)
 		}
 	}
-}
-
-func TestEntropyMnemonicSeed(t *testing.T) {
-	// Do the whole rigamarole of Entropy -> Mnemonic -> Seed
-	entropy, err := bip39.NewEntropy(128) // 128 bits of entropy
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(entropy))
-
-	mnemonic, err := bip39.NewMnemonic(entropy)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(mnemonic))
-
-	entropyX, err := bip39.EntropyFromMnemonic(mnemonic)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(entropyX))
-
-	if !bytes.Equal(entropy, entropyX) {
-		t.Fatal("entropy not equal")
-	}
-
-	// The following is educational only on how to get to a master key.
-	seed, err := bip39.NewSeedWithErrorChecking(mnemonic, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(seed))
-
-	seed2 := bip39.NewSeed(mnemonic, "")
-	t.Log(spew.Sdump(seed2))
-	if !bytes.Equal(seed, seed2) {
-		t.Fatal("seeds not equal")
-	}
-
-	mk, err := hdkeychain.NewMaster(seed, &chaincfg.MainNetParams)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(mk)) // check on https://iancoleman.io/bip39
-}
-
-func TestDeriveAddresses(t *testing.T) {
-	// Test vectors generated on https://iancoleman.io/bip39 using bip32
-	mnemonic := "dinosaur banner version pistol need area dream champion kiss thank business shrug explain intact puzzle"
-
-	seed, err := bip39.NewSeedWithErrorChecking(mnemonic, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	mk, err := hdkeychain.NewMaster(seed, &chaincfg.MainNetParams)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(mk)) // check on https://iancoleman.io/bip39
-
-	// Derive extended key for hardened account 0: m/0'
-	acct0, err := mk.Derive(hdkeychain.HardenedKeyStart + 0)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(acct0))
-
-	// Derive extended key for external hardened account 0 m/0'/0'
-	acct0eh, err := acct0.Derive(hdkeychain.HardenedKeyStart + 0)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(acct0eh))
-	pkHHash, err := acct0eh.Address(&chaincfg.MainNetParams)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(pkHHash))
-	if pkHHash.String() != "14fjqmZJU7qrRtWHgVjh6jcoKjqu6y1gyM" {
-		t.Fatal("invalid hardned address")
-	}
-
-	// Derive extended key for external account 0 m/0'/0
-	acct0e, err := acct0.Derive(0)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(acct0e))
-	pkHash, err := acct0e.Address(&chaincfg.MainNetParams)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(pkHash))
-	if pkHash.String() != "1KWtnYZmQwnwCdGNwjbzpvh3YibNV6KWb9" {
-		t.Fatal("invalid address")
-	}
-}
-
-func TestWalletCreate(t *testing.T) {
-	mnemonic := "dinosaur banner version pistol need area dream champion kiss thank business shrug explain intact puzzle"
-	seed := "5e2deaa9f1bb2bcef294cc36513c591c5594d6b671fe83a104aa2708bc634cb0602599b867332dfec245547baafae40dad247f21564a0de925527f2445a086fd"
-	xpriv := "xprv9s21ZrQH143K4H14eaYTxfuxn6eDqoPhuQx2sRr9ZvmvkBQ39KMHfaNQ9GDdUkgRbibGXu66XQTZj9QJRSXBEibqrHb34BdJRPvqqiZMTCA"
-	expectedAddress := "14fjqmZJU7qrRtWHgVjh6jcoKjqu6y1gyM"
-	expectedPub, _ := hex.DecodeString("02dc2f5439d4e4e1d7da99daa5707d7d7da72caf4e31c5aa206322e00bc1f2ce8c")
-
-	w, err := WalletNew(&chaincfg.MainNetParams)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	for k, v := range []string{mnemonic, seed, "0x" + seed, xpriv} {
-		t.Logf("using secret: %v", v)
-		err = w.Unlock(v)
-		if err != nil {
-			t.Fatalf("failed %v: %v", k, err)
-		}
-
-		ek, err := w.DeriveHD(0, 0)
-		if err != nil {
-			t.Fatal(err)
-		}
-		addr, pub, err := AddressAndPublicFromExtended(&chaincfg.MainNetParams, ek)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if addr.String() != expectedAddress {
-			t.Fatal("invalid hardened address")
-		}
-		expub, err := Compressed(pub)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if !bytes.Equal(expub, expectedPub) {
-			t.Fatal("invalid hardened compressed public key")
-		}
-		err = w.Lock()
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// use a pub key to test negative path
-	failpub := "xpub68GbL9kRsJSW9YgqXTnAy1JsiCjZEwbecAx1dcNhxdypbFmsDP7NviUgvWmX7ighnsTe9eLJQHDcnoGQUtVKViSnAKzxQ2DHC6QeNbYiztE"
-	err = w.Unlock(failpub)
-	if err == nil {
-		t.Fatal("expected error")
-	}
-}
-
-// XXX duplicate code
-func executeTX(t *testing.T, dump bool, scriptPubKey []byte, tx *btcutil.Tx) error {
-	flags := txscript.ScriptBip16 | txscript.ScriptVerifyDERSignatures |
-		txscript.ScriptStrictMultiSig | txscript.ScriptDiscourageUpgradableNops
-	vm, err := txscript.NewEngine(scriptPubKey, tx.MsgTx(), 0, flags, nil, nil, -1, nil)
-	if err != nil {
-		return err
-	}
-	if dump {
-		t.Logf("=== executing tx %v", tx.Hash())
-	}
-	for i := 0; ; i++ {
-		d, err := vm.DisasmPC()
-		if dump {
-			t.Logf("%v: %v", i, d)
-		}
-		done, err := vm.Step()
-		if err != nil {
-			return err
-		}
-		stack := vm.GetStack()
-		if dump {
-			t.Logf("%v: stack %v", i, spew.Sdump(stack))
-		}
-		if done {
-			break
-		}
-	}
-	err = vm.CheckErrorCondition(true)
-	if err != nil {
-		return err
-	}
-
-	if dump {
-		t.Logf("=== SUCCESS tx %v", tx.Hash())
-	}
-	return nil
-}
-
-func TestTransactionCreate(t *testing.T) {
-	// KeyStore for key looksups during signing
-	ks, err := memoryKeyStoreNew(&chaincfg.TestNet3Params)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	mnemonic := "dinosaur banner version pistol need area dream champion kiss thank business shrug explain intact puzzle"
-	w, err := WalletNew(&chaincfg.TestNet3Params)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = w.Unlock(mnemonic)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ek, err := w.DeriveHD(0, 0)
-	if err != nil {
-		t.Fatal(err)
-	}
-	addr, pub, err := AddressAndPublicFromExtended(&chaincfg.TestNet3Params, ek)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("%v", addr)
-	t.Logf("%v", pub)
-
-	// Store in key store
-	err = ks.Put(&NamedKey{
-		Name:       "my private key",
-		Account:    0,
-		Child:      0,
-		HD:         true,
-		PrivateKey: ek,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	pkscript, err := ScriptFromPubKeyHash(addr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("%x", pkscript)
-	scripthash := ScriptHashFromScript(pkscript)
-	t.Logf("%v", scripthash)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	b, err := BlockstreamNew(&chaincfg.TestNet3Params)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	feeEstimates, err := b.FeeEstimates(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	feeEstimate, err := FeeByConfirmations(6, feeEstimates)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(feeEstimate))
-
-	utxos, err := b.UtxosByAddress(ctx, addr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("balance %v: %v", addr, BalanceFromUtxos(utxos))
-
-	//// pick utxo
-	//amount := btcutil.Amount(1000000) // 0.01000000 BTC
-	//fee := btcutil.Amount(50000)      // 0.00050000 BTC
-	//total := amount + fee             // 0.01050000 BTC
-	//utxo, err := UtxoPickerSingle(amount, fee, utxos)
-	//if err != nil {
-	//	t.Fatal(err)
-	//}
-	//t.Logf("utxo: %v > %v", btcutil.Amount(utxo.Value), total)
-
-	keystone := &hemi.L2Keystone{
-		Version:            1,
-		L1BlockNumber:      0xbadc0ffe,
-		L2BlockNumber:      0xdeadbeef,
-		ParentEPHash:       digest256([]byte{1, 1, 3, 7}),
-		PrevKeystoneEPHash: digest256([]byte{0x04, 0x20, 69}),
-		StateRoot:          digest256([]byte("Hello, world!")),
-		EPHash:             digest256([]byte{0xaa, 0x55}),
-	}
-	tx, prevOut, err := PoPTransactionCreate(keystone, uint32(time.Now().Unix()),
-		btcutil.Amount(feeEstimate.SatsPerByte+0.5), utxos, pkscript)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	err = TransactionSign(&chaincfg.TestNet3Params, ks, tx, prevOut)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("tx: %v", spew.Sdump(tx))
-
-	err = executeTX(t, true, tx.TxOut[0].PkScript, btcutil.NewTx(tx))
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestTBCWallet(t *testing.T) {
-	mnemonic := "dinosaur banner version pistol need area dream champion kiss thank business shrug explain intact puzzle"
-
-	w, err := WalletNew(&chaincfg.TestNet3Params)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = w.Unlock(mnemonic)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ek, err := w.DeriveHD(0, 0)
-	if err != nil {
-		t.Fatal(err)
-	}
-	addr, pub, err := AddressAndPublicFromExtended(&chaincfg.TestNet3Params, ek)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("%v", addr)
-	t.Logf("%v", pub)
-
-	pkscript, err := ScriptFromPubKeyHash(addr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("%x", pkscript)
-	scripthash := ScriptHashFromScript(pkscript)
-	t.Logf("%v", scripthash)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-
-	// Connect tbc service
-	tbcCfg := &tbc.Config{
-		AutoIndex:            false,
-		BlockCacheSize:       "10mb",
-		BlockheaderCacheSize: "1mb",
-		BlockSanity:          false,
-		LevelDBHome:          t.TempDir(),
-		// LogLevel:                "tbcd=TRACE:tbc=TRACE:level=DEBUG",
-		MaxCachedTxs:            1000, // XXX
-		Network:                 "localnet",
-		PrometheusListenAddress: "",
-		Seeds:                   []string{"127.0.0.1:18444"},
-		ListenAddress:           "localhost:8881",
-	}
-	_ = loggo.ConfigureLoggers(tbcCfg.LogLevel)
-	s, err := tbc.NewServer(tbcCfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-	go func() {
-		err := s.Run(ctx)
-		if err != nil && !errors.Is(err, context.Canceled) && !errors.Is(err, rawpeer.ErrNoConn) {
-			panic(err)
-		}
-	}()
-
-	cfg := &Config{
-		TBCWSURL:     fmt.Sprintf("ws://%s/%s", tbcCfg.ListenAddress, tbcapi.RouteWebsocket),
-		BTCChainName: "testnet3",
-	}
-
-	time.Sleep(1 * time.Second)
-
-	defer cancel()
-	b, err := TBCNodeNew(ctx, cfg, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	time.Sleep(1 * time.Second)
-
-	feeEstimates, err := b.FeeEstimates(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	feeEstimate, err := FeeByConfirmations(6, feeEstimates)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(spew.Sdump(feeEstimate))
-
-	utxos, err := b.UtxosByAddress(ctx, addr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("balance %v: %v", addr, BalanceFromUtxos(utxos))
-
-	/* // pick utxo
-	amount := btcutil.Amount(1000000) // 0.01000000 BTC
-	fee := btcutil.Amount(50000)      // 0.00050000 BTC
-	total := amount + fee             // 0.01050000 BTC
-	utxo, err := UtxoPickerSingle(amount, fee, utxos)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("utxo: %v > %v", btcutil.Amount(utxo.Value), total) */
 }
