+++ bitcoin/wallet/README.md
@@ -1,0 +1,29 @@
+# wallet
+
+Wallet is a collection of convenience structures, functions and interfaces that
+can be tied together to create a bitcoin wallet. All wallets have slightly
+different requirements and these packages try to break a wallet down into its
+constituent components.
+
+The components are:
+* gozer - Interface to a bitcoin source of truth.
+* vinzclortho - Package that handles creation and derivation bitcoin addresses.
+* zuul - Package that handles storage of secret material. 
+
+## Canon
+
+Gozer was originally worshiped as a god by the Hittites, Mesopotamians, and the
+Sumerians around 6000 BC. Gozer was genderless and could assume any form it
+wanted. Gozer had two trusted minions – themselves worshiped as demigods – that
+were harbingers of destruction and primary agents for its arrival: Vinz
+Clortho, The Keymaster and Zuul, The Gatekeeper. The Sumerians believed in a
+land of the dead, a dark and shadowy realm within the bowels of the earth where
+the souls of the dead were ruled by Gozer and protected by the Gatekeeper and
+Keymaster. Gozer also consumed the souls of those sacrificed to it. In order
+for Gozer to rise up and walk the human plane again, the Gatekeeper and
+Keymaster had to assume the form of beasts. During the rectification of
+the Vuldronaii, Gozer came as a large moving Torb. Then, during the third
+reconciliation of the last of the Meketrex Supplicants, they chose the form of
+a giant Sloar.
+
+Source https://ghostbusters.fandom.com/wiki/Gozer
