+++ bitcoin/wallet/gozer/tbc/tbc.go
@@ -1,0 +1,1 @@
+package tbc
